Refactor gcsfs/_dircache.py
