+++ gcsfs/extended_gcsfs.py
@@ -26,6 +26,7 @@ from google.cloud.storage.asyncio.async_multi_range_downloader import (
 
 from gcsfs import __version__ as version
 from gcsfs import zb_hns_utils
+from gcsfs._dircache import HnsDirCacheUpdater
 from gcsfs.core import GCSFile, GCSFileSystem
 from gcsfs.retry import DEFAULT_RETRY_CONFIG, get_storage_control_retry_config
 from gcsfs.zb_hns_utils import DirectMemmoveBuffer, MRDPool
@@ -77,7 +78,7 @@ async def _get_mrd_size(mrd_or_pool):
         return m.persisted_size
 
 
-class ExtendedGcsFileSystem(GCSFileSystem):
+class ExtendedGcsFileSystem(HnsDirCacheUpdater, GCSFileSystem):
     """
     This class will be used when GCSFS_EXPERIMENTAL_ZB_HNS_SUPPORT env variable is set to true.
     ExtendedGcsFileSystem is a subclass of GCSFileSystem that adds new logic for bucket types
@@ -632,164 +633,6 @@ class ExtendedGcsFileSystem(GCSFileSystem):
 
         return bucket_type in [BucketType.ZONAL_HIERARCHICAL, BucketType.HIERARCHICAL]
 
-    def _cache_drop_entries(self, parent, names):
-        """Remove entries whose ``name`` is in ``names`` from ``parent``'s
-        cached listing. No-op when ``parent`` is not cached.
-
-        ``names`` is a set of stripped ``bucket/key`` names matching the
-        ``name`` field of the cached entries.
-        """
-        if parent in self.dircache:
-            self.dircache[parent] = [
-                e for e in self.dircache[parent] if e.get("name") not in names
-            ]
-
-    def _cache_add_entry(self, parent, entry):
-        """Append ``entry`` to ``parent``'s cached listing. No-op when
-        ``parent`` is not cached."""
-        if parent in self.dircache:
-            self.dircache[parent].append(entry)
-
-    def _update_dircache_after_rename(self, path1, path2):
-        """
-        Performs a targeted update of the directory cache after a successful
-        folder rename operation.
-
-        This involves three main steps:
-        1. Removing the source folder and all its descendants from the cache.
-        2. Removing the source folder's entry from its parent's listing.
-        3. Adding the new destination folder's entry to its parent's listing.
-
-        Args:
-            path1 (str): The source path that was renamed.
-            path2 (str): The destination path.
-        """
-        # 1. Find and remove all descendant paths of the source from the cache.
-        source_prefix = f"{path1.rstrip('/')}/"
-        for key in list(self.dircache):
-            if key.startswith(source_prefix):
-                self.dircache.pop(key, None)
-
-        # 2. Remove the old source entry from its parent's listing.
-        self.dircache.pop(path1, None)
-        self._cache_drop_entries(self._parent(path1), {path1})
-
-        # 3. Invalidate the destination path and update its parent's cache.
-        self.dircache.pop(path2, None)
-        _, key2, _ = self.split_path(path2)
-        new_entry = {
-            "Key": key2,
-            "Size": 0,
-            "name": path2,
-            "size": 0,
-            "type": "directory",
-            "storageClass": "DIRECTORY",
-        }
-        self._cache_add_entry(self._parent(path2), new_entry)
-
-    async def _mv_file_cache_update(self, path1, path2, response=None):
-        """
-        Update the cache after a file move operation.
-
-        For HNS-enabled buckets where the move is within the same bucket, this method
-        directly updates the directory cache by removing the source entry from it's
-        parent cache and adding destination path as a new entry in it's corresponding parent cache.
-        This avoids invalidating the entire parent directory cache, which is beneficial for HNS
-        performance.
-
-        For non-HNS buckets or cross-bucket moves, it falls back to the default
-        behavior (invalidating the cache for both source and destination parents).
-        """
-        src_bucket, _, _ = self.split_path(path1)
-        dest_bucket, _, _ = self.split_path(path2)
-
-        if await self._is_bucket_hns_enabled(src_bucket) and src_bucket == dest_bucket:
-            # Source: removing the entry never changes an ancestor's listing, so
-            # drop it in place.
-            self._cache_drop_entries(self._parent(path1), {self._strip_protocol(path1)})
-            dest_parent = self._parent(path2)
-            if response and dest_parent in self.dircache:
-                # Destination parent is already cached (so it existed before the
-                # move) -> updating it in place cannot leave an ancestor stale.
-                self._cache_add_entry(
-                    dest_parent, self._process_object(dest_bucket, response)
-                )
-            else:
-                # The destination parent may have been created by this move;
-                # broad-invalidate it so a cached ancestor can't hide the new
-                # directory (also covers a missing/empty move response).
-                self.invalidate_cache(dest_parent)
-        else:
-            await super()._mv_file_cache_update(path1, path2, response)
-
-    async def _rm_files_cache_update(self, paths):
-        """
-        Update the cache after a (batch) file delete operation.
-
-        For HNS-enabled buckets, directories are first-class persistent objects,
-        so deleting a file only changes the contents of its immediate parent.
-        Each deleted entry is removed from its immediate parent's listing
-        (grouped per parent so each listing is rewritten only once), avoiding the
-        redundant invalidation of all ancestor directories up to the root.
-
-        Entries are matched by name only; the object generation is not taken
-        into account. Non-HNS paths fall back to the default broad invalidation
-        of the parents and all of their ancestors.
-        """
-        # Split each path once and resolve each distinct bucket's HNS status
-        # once, concurrently, rather than awaiting a (cached) lookup per path.
-        split_paths = [(path, *self.split_path(path)) for path in paths]
-        buckets = list({bucket for _, bucket, _, _ in split_paths})
-        hns_enabled = dict(
-            zip(
-                buckets,
-                await asyncio.gather(
-                    *(self._is_bucket_hns_enabled(bucket) for bucket in buckets)
-                ),
-            )
-        )
-
-        # Group the names to drop by their immediate parent so each listing is
-        # rewritten only once. Non-HNS paths fall back to broad invalidation.
-        removed_names_by_parent = {}
-        non_hns_paths = []
-        for path, bucket, key, _ in split_paths:
-            if hns_enabled[bucket]:
-                removed_names_by_parent.setdefault(self._parent(path), set()).add(
-                    f"{bucket}/{key}"
-                )
-            else:
-                non_hns_paths.append(path)
-
-        for parent, removed_names in removed_names_by_parent.items():
-            self._cache_drop_entries(parent, removed_names)
-
-        if non_hns_paths:
-            await super()._rm_files_cache_update(non_hns_paths)
-
-    async def _write_file_cache_update(self, path):
-        """
-        Update the cache after a file create/overwrite (put/pipe/cp).
-
-        For HNS-enabled buckets the single-level shortcut (invalidate only the
-        immediate parent, not every ancestor up to the bucket root) is taken
-        only when that parent is already cached. A cached parent is one we have
-        listed, so it already exists, which means creating a file inside it
-        cannot add a new directory to any ancestor's listing.
-
-        When the parent is not cached we cannot rule out that this write
-        implicitly created it (HNS auto-creates missing parent folders on object
-        write); leaving a cached ancestor untouched would hide the new directory
-        from a subsequent listing. In that case, and for non-HNS buckets, fall
-        back to the default broad invalidation of the parent and all ancestors.
-        """
-        bucket, _, _ = self.split_path(path)
-        parent = self._parent(path)
-        if await self._is_bucket_hns_enabled(bucket) and parent in self.dircache:
-            self.dircache.pop(parent, None)
-        else:
-            await super()._write_file_cache_update(path)
-
     async def _mv(self, path1, path2, **kwargs):
         """
         Move a file or directory. Overrides the parent `_mv` to provide an
@@ -1018,15 +861,10 @@ class ExtendedGcsFileSystem(GCSFileSystem):
                 timeout=STORAGE_CONTROL_RPC_TIMEOUT,
             )
             # Instead of invalidating the parent cache, update it to add the new entry.
-            new_entry = {
-                "Key": key.rstrip("/"),
-                "Size": 0,
-                "name": path,
-                "size": 0,
-                "type": "directory",
-                "storageClass": "DIRECTORY",
-            }
-            self._cache_add_entry(self._parent(path), new_entry)
+            self._cache_add_entry(
+                self._parent(path),
+                self._directory_cache_entry(path, key.rstrip("/")),
+            )
         except api_exceptions.Conflict as e:
             logger.debug(f"Directory already exists: {path}: {e}")
         except api_exceptions.FailedPrecondition as e:
